+++ pkg/metric_generator/generator_test.go
@@ -21,8 +21,8 @@ import (
 	"sort"
 	"testing"
 
-	"k8s.io/kube-state-metrics/pkg/allow"
-	"k8s.io/kube-state-metrics/pkg/metric"
+	"k8s.io/kube-state-metrics/v2/pkg/allow"
+	"k8s.io/kube-state-metrics/v2/pkg/metric"
 )
 
 func TestFilterMetricFamiliesLabels(t *testing.T) {